Finally remove libpinyin-tools, merge to libpinyin0-dev.
These three tools aren't designed for ordinary users to use, they need a
fixed sequence to be executed, otherwise the latter two will fail as with
sigmentation faults. This is not expected normally, and may result into
many bug reports. I'm installing it in /usr/lib/libpinyin/tools, together
with a detailed README.Debian, and wish this helps.